feat: Add htmltest to the documenation build
Keeping links in documentation up to date is hard, sometimes links
are breaking, sometimes things will get moved around and the links
are not up-to-date anymore. HTMLTests checks for certain things,
and helps to keep the documentation up-to-date.

Additionally i added a makefile using docker to reduce the need of
local dependencies and create a unified build experience for CI
and local systems.

Signed-off-by: Simon Schrottner <simon.schrottner@dynatrace.com>